Refactor GroupsList page
