+++ app/src/pages/GroupsList/GroupsList.jsx
@@ -1,111 +1,34 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { Link } from 'react-router-dom';
-import { Helmet } from 'react-helmet';
-import { debounce } from 'lodash';
+import React, { useState, useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import { groupActions, groupSelectors } from 'redux/groups';
-import { PageTitle, VerifiedBadge, TextInput, TextButton, TablePlaceholder, Table } from 'components';
+import { debounce } from 'lodash';
+import { Helmet } from 'react-helmet';
+import { useLazyLoading } from 'hooks';
+import { PageTitle, TextButton, TextInput } from 'components';
+import { List } from './containers';
 import './GroupsList.scss';
-
-const RESULTS_PER_PAGE = 20;
-
-const TABLE_CONFIG = {
-  uniqueKey: row => row.id,
-  columns: [
-    {
-      key: 'name',
-      className: () => '-primary',
-      transform: (val, row) => {
-        if (row.verified) {
-          return (
-            <Link to={`/groups/${row.id}`}>
-              <VerifiedBadge />
-              {val}
-            </Link>
-          );
-        }
-
-        return <Link to={`/groups/${row.id}`}>{val}</Link>;
-      }
-    },
-    {
-      key: 'memberCount',
-      className: () => '-break-small',
-      transform: val => `${val} members`,
-      width: 130
-    }
-  ]
-};
 
 function GroupsList() {
   const dispatch = useDispatch();
 
   // State variables
   const [nameSearch, setNameSearch] = useState('');
-  const [pageIndex, setPageIndex] = useState(0);
 
-  // Memoized redux variables
-  const groups = useSelector(groupSelectors.getGroups);
-  const isLoading = useSelector(groupSelectors.isFetchingAll);
+  const { isFullyLoaded, reloadData } = useLazyLoading({
+    resultsPerPage: 20,
+    selector: groupSelectors.getGroups,
+    action: handleLoadData
+  });
 
-  const isFullyLoaded = groups.length < RESULTS_PER_PAGE * (pageIndex + 1);
-
-  const handleSubmitSearch = debounce(
-    () => {
-      setPageIndex(0); // Reset pagination when the search changes
-      dispatch(groupActions.fetchList(nameSearch, RESULTS_PER_PAGE, 0));
-    },
-    500,
-    { leading: true, trailing: false }
-  );
-
-  const handleLoadMore = () => {
-    if (pageIndex === 0) return;
-
-    const limit = RESULTS_PER_PAGE;
-    const offset = RESULTS_PER_PAGE * pageIndex;
+  function handleLoadData(limit, offset) {
     dispatch(groupActions.fetchList(nameSearch, limit, offset));
-  };
+  }
 
-  const handleNextPage = () => {
-    setPageIndex(pageIndex + 1);
-  };
-
-  const handleNameSearchInput = e => {
-    setNameSearch(e.target.value);
-  };
-
-  const handleScrolling = () => {
-    const margin = 300;
-
-    window.onscroll = debounce(() => {
-      // If has no more content to load, ignore the scrolling
-      if (groups.length < RESULTS_PER_PAGE * (pageIndex + 1)) {
-        return;
-      }
-
-      const { innerHeight } = window;
-      const { scrollTop, offsetHeight } = document.documentElement;
-
-      // If has reached the bottom of the page, load more data
-      if (innerHeight + scrollTop + margin > offsetHeight) {
-        // This timeout is simply to wait for the scrolling
-        // inertia to stop, to then load the contents, otherwise
-        // it will resume the inertia and scroll to the bottom of the page again
-        setTimeout(() => handleNextPage(), 800);
-      }
-    }, 100);
-  };
-
-  // Memoized callbacks
-  const onSubmitSearch = useCallback(handleSubmitSearch, [nameSearch]);
-  const onLoadMore = useCallback(handleLoadMore, [pageIndex]);
-  const onNameSearchInput = useCallback(handleNameSearchInput, [setNameSearch]);
+  // Debounce search input keystrokes by 500ms
+  const handleSubmitSearch = debounce(reloadData, 500, { leading: true, trailing: false });
 
   // Submit search each time any of the search variable change
-  useEffect(onLoadMore, [pageIndex]);
-  useEffect(handleScrolling, [groups, pageIndex]);
-  useEffect(onSubmitSearch, [nameSearch]);
+  useEffect(handleSubmitSearch, [nameSearch]);
 
   return (
     <div className="groups__container container">
@@ -122,31 +45,14 @@ function GroupsList() {
       </div>
       <div className="groups__options row">
         <div className="col-12">
-          <TextInput onChange={onNameSearchInput} placeholder="Search group" />
+          <TextInput onChange={e => setNameSearch(e.target.value)} placeholder="Search competition" />
         </div>
       </div>
       <div className="groups__list row">
-        <div className="col">
-          {isLoading && (!groups || groups.length === 0) ? (
-            <TablePlaceholder size={5} />
-          ) : (
-            <Table
-              uniqueKeySelector={TABLE_CONFIG.uniqueKey}
-              columns={TABLE_CONFIG.columns}
-              rows={groups}
-              listStyle
-            />
-          )}
-        </div>
+        <List />
       </div>
       <div className="row">
-        <div className="col">
-          {!isFullyLoaded && (
-            <b id="loading" className="loading-indicator">
-              Loading...
-            </b>
-          )}
-        </div>
+        <div className="col">{!isFullyLoaded && <b className="loading-indicator">Loading...</b>}</div>
       </div>
     </div>
   );
